Added MASE Vanilla version
Splits the MASE metric into two versions, 

MASE_vanilla: 
This version is true to the MASE definition, thus allowing division by zero in certain cases. 

MASE:
Puts a lower bound on the denominator value when dividing allowing fixing the previous discontinuity

A good example of the effect of these metrics is to run backtest_metrics on the constant dataset with any estimator. 

The previous version of the metric returned faulty values in this case.