Use `pd.Period` instead of `pd.Timestamp`. (#1980)
In pandas, a `Timestamp` is just a point in time with nanosecond resolution
using 64bits, i.e. a numpy `datetime64[ns]`. There are obvious downsides using
such a high precision when the finest granularity that we support is one
minute. For one, it only supports timestamps between the years ~1750 and ~2250.
Second, computation is more complex; for example adding a year to a timestamp
takes 10x than using a `pandas.Period` with yearly frequency.

Further, a `Period` always represents a time-span and has a frequency, while
pandas will drop frequency support for timestamps.

In short, `Period` and `period_index` is what we actually wanted to use from
the start.