Fix numerical bug
prob.cumsum() - prob is a numerically unsafe way to shift prob.cumsum() to the right. The numerical error will create small gaps or overlaps in the bins defined by lower and upper bounds. If a quantile falls in one of these overlap or gap the one hot bin indicator will have 0 or 2 hot bins and generate a shape mismatch error. The fix consists of doing a shift by copy.