Backports for v0.10.8 (#2400)
* Fix numerical bug in `BinnedUniforms` (#2344)

prob.cumsum() - prob is a numerically unsafe way to shift prob.cumsum() to the right. The numerical error will create small gaps or overlaps in the bins defined by lower and upper bounds. If a quantile falls in one of these overlap or gap the one hot bin indicator will have 0 or 2 hot bins and generate a shape mismatch error. The fix consists of doing a shift by copy.

* Fix dominick dataset bug. (#2364)

* Proposed fix to zero seed bug. (#2379)

Co-authored-by: Sigrid Passano Hellan <sphellan@amazon.de>

Co-authored-by: Marc van Oudheusden <60158499+moudheus@users.noreply.github.com>
Co-authored-by: Bhaskar Dhariyal <20501023+haskarb@users.noreply.github.com>
Co-authored-by: Sigrid Passano Hellan <sighellan@gmail.com>
Co-authored-by: Sigrid Passano Hellan <sphellan@amazon.de>