Recover from a checkpoint if training failed with an exception (#3265)
*Issue #, if available:*

*Description of changes:*
- When training models that rely on the `StundentTOutput` (e.g.,
PatchTST, DeepAR), we occasionally run into numerical issues that turn
all models weights to NaNs in the middle of the training:
    ```
ValueError: Expected parameter df (Tensor of shape (64, 179)) of
distribution Chi2() to satisfy the constraint
GreaterThan(lower_bound=0.0), but found invalid values:
    tensor([[nan, nan, nan,  ..., nan, nan, nan],
            [nan, nan, nan,  ..., nan, nan, nan],
            [nan, nan, nan,  ..., nan, nan, nan],
            ...,
            [nan, nan, nan,  ..., nan, nan, nan],
            [nan, nan, nan,  ..., nan, nan, nan],
            [nan, nan, nan,  ..., nan, nan, nan]])
    ```
    Currently, this completely breaks the training process. 
    
With this PR, we attempt to recover the model weights from a checkpoint
if an exception is encountered during training.
 
A potential alternative to this PR would be to fix the numerical issues
plaguing the StudentT distribution by, e.g., clipping parameters like
the scale. However, this might result in regressions for some existing
users. Also, it's not very clear when and why this problem occurs, so
even if we introduce clipping now, there might still be other datasets
where it won't be sufficient.
- Bump `lightning` upper bound to `<2.6`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.


**Please tag this pr with at least one of these labels to make our
release process faster:** BREAKING, new feature, bug fix, other change,
dev setup